app/requirement.txt
cat > requirements.txt <<'EOF'
fastapi==0.95.2
uvicorn[standard]==0.22.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.2.2
pmdarima==2.2.4
prophet==1.1.2
statsmodels==0.14.0
streamlit==1.28.0
mlflow==2.5.1
psycopg2-binary==2.9.7
duckdb==0.8.1
boto3==1.28.6
pyyaml==6.0
shap==0.42.1
pytest==8.4.2
python-dotenv==1.0.0
loguru==0.7.0
EOF
